Phrases checks
Now offers corrections for phrases as well, with indication of which word in the phrase the mistake was made, as well as the position of letter in the word that was wrong.

I also updated the wordorphrase var, using Math.floor was causing the wordorphrase variable to never be greater than 1.